Finish the feature

+++ src/shading.cpp
@@ -47,45 +47,6 @@ const glm::vec3 computeShading(const glm::vec3& lightPosition, const glm::vec3& 
     glm::vec3 viewVec = cameraToPoint - ray.origin;
     glm::vec3 lightVec = lightPosition - cameraToPoint;
     glm::vec3 reflected = normalize(-normalize(lightVec) + 2 * (glm::dot(normalize(lightVec), hitInfo.normal)) * hitInfo.normal);
-    /*
-    * GLOSSY REFLECTIONS
-    */
-    //if (features.extra.enableGlossyReflection) {
-    //    float a = 1 / ((hitInfo.material.shininess) * (hitInfo.material.shininess));
-
-    //    drawRay({ cameraToPoint, reflected, 0.1f }, { 0.0f, 0.0f, 1.0f });
-
-    //    // 1. Find the orthonormal basis:
-
-    //    glm::vec3 w_vec = glm::normalize(reflected);
-    //    // u_vec is supposed to be not collinear with w: to find it
-    //    // we set the smallest u_vec component to 1 (2.4.6 book)
-    //    glm::vec3 u_vec = w_vec;
-    //    if (u_vec.x < u_vec.y && u_vec.x < u_vec.z)
-    //        u_vec.x = 1;
-    //    else if (u_vec.y < u_vec.x && u_vec.y < u_vec.z)
-    //        u_vec.y = 1;
-    //    else
-    //        u_vec.z = 1;
-    //    u_vec = glm::normalize(u_vec);
-
-    //    glm::vec3 v_vec = glm::normalize(glm::cross(w_vec, u_vec));
-
-    //    // 2. Create 2 random points in [0,1]:
-    //    float ran1 = Generator::get(0.0f, 1.0f);
-    //    float ran2 = Generator::get(0.0f, 1.0f);
-
-    //    // 3. Calculate vector coefficients:
-    //    float u = -a / 2 + ran1 * a;
-    //    float v = -a / 2 + ran2 * a;
-
-    //    // 4. Replace reflected with a perturbed reflected vector
-    //    reflected = reflected + u * u_vec + v * v_vec;
-    //    std::cout << "u: " << u << ", v: " << v << '\n';
-
-    //    drawRay({ cameraToPoint, reflected, 0.1f }, { 1.0f, 0.0f, 1.0f });
-    //}
-  
     glm::vec3 specular = lightColor * hitInfo.material.ks * pow(glm::max(0.0f, glm::dot(normalize(viewVec), reflected)), hitInfo.material.shininess);
 
     return diffuse + specular;
@@ -149,5 +110,5 @@ glm::vec3 distortRayDirection(const Ray& ray, float shininess, int samples)
     }
 
     // 5. Return distorted direction
-    return direction / glm::vec3 { samples };
+    return direction / glm::vec3 { (float)samples };
 }
